requirements.txt
pyttsx3
python-dotenv==1.0.1
huggingface-hub==0.26.5
speechbrain==1.0.2
sounddevice
wavio
pygame
torchaudio
torch
transformers
noisereduce==3.0.3
librosa==0.10.2.post1
soundfile
numpy
playsound==1.3.0
PyGObject==3.50.0
